fix setting the HOME variable to contain the expected value
Using the SetEnvironmentVariable method does not work because it will be added
as literal and not correctly expanded (despite the Windows Environment Settings
dialog showing it as such) in e.g. Cmd or Powershell.

This happens because values set via this method have the type REG_SZ. The
appropriate type when using recognized environment variables is REG_EXPAND_SZ
to force expansion on use in all cases.

See https://stackoverflow.com/a/11295260/2920585 for additional details.